requirements.txt
pytesseract>=0.3.10
opencv-python>=4.5.0.0
Pillow>=8.0.0
numpy>=1.19.0